TaskTracker 1.0
This application allows the user to read tasks from a JSON file, add, update and delete from it.

The application implements a listener/subscriber functionality to automatically update the list with changes that are reflected in the "list" commands.

This project was made in order to practice using interfaces, factory - repository - model system design methodologies, enums and 3rd party libaries to manipulate JSONObject in java.